requirements.txt
# Core Dependencies
django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
python-decouple==3.8

# FastAPI & Async
fastapi==0.115.0
uvicorn[standard]==0.32.0
gunicorn==23.0.0
hypercorn==0.17.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0

# Redis & Caching
redis==5.1.1
django-redis==5.4.0
hiredis==2.3.2

# Pydantic & Validation
pydantic==2.9.2
pydantic-settings==2.6.1
email-validator==2.2.0

# Background Tasks
celery==5.4.0
celery[redis]==5.4.0
kombu==5.4.2

# Testing
pytest==8.3.3
pytest-django==4.9.0
pytest-asyncio==0.24.0
pytest-cov==6.0.0
factory-boy==3.3.1
faker==30.8.1
httpx==0.28.1

# Development & Code Quality
black==24.10.0
isort==5.13.2
flake8==7.1.1
mypy==1.12.1
django-stubs==5.1.1
djangorestframework-stubs==3.15.2
pre-commit==4.0.1

# Security & Authentication
cryptography==43.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bleach==6.2.0

# Monitoring & Logging
structlog==24.4.0
sentry-sdk==2.17.0
prometheus-client==0.20.0

# Documentation
drf-spectacular==0.27.2

# Performance & Optimization
django-debug-toolbar==4.4.6
django-silk==5.1.0
orjson==3.10.12

# File Storage & Processing
Pillow==10.4.0
python-magic==0.4.27
boto3==1.35.52

# API & Serialization
requests==2.32.3
aiohttp==3.11.2
msgspec==0.18.6